+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.2.0
 * super fast tooltips for large graphs
 * smart label zooming
+* group focusing on node click, or via graph.selectGroup()
 * upgraded to webgl2 and GLSL v3.0
 
 ## v2.1.1